Publish root document
Commit 07060fd started only publishing the `compiler-user-guide`
subdirectory, but that means we no longer upload the root document that
redirects us there.

We need to set `keep_files` to not delete the PR previews. This is not
ideal, since it will also make any intended removal a manual task.

Let's also make sure the CNAME file is there. With `keep_files` set, it
will generally stay where it needs to be, but including it every time is
more robust.

it was already there and the cleanup of `actions-gh-pages` only cleans
up inside the `destination_dir`.