Fetch PR previews when deploying (#13)
When constructing the new site deployment, extract from the current
`gh-pages` branch an explicit subset of the files, namely all
directories whose names start with `pr-`. In effect, we keep those
directories as they were.

We want the build of the website to retain the PR previews we generate
in subdirectories called `pr-<number>/`. The configuration knobs provided by
`peaceiris/actions-gh-pages` don't give us a solution.

Before this commit, we restricted the deployment to a subdirectory,
which will clean up only the files in that subdirectory and leave the
rest alone. But this does not work because we want multiple
subdirectories; currently we have `compiler-user-guide`, but soon we
will add `tutorial`. It also means we would need to maintain the root
`index.html` separately from the `main` branch where we maintain other
files: we would need to maintain `index.html` directly in the `gh-pages`
branch, which is undesirable.

We could assert `keep_files`, but this is too blunt an instrument: any
file ever published will live on until manually deleted.

Therefore, we do not use either of these possibilities provided by
`peaceiris/actions-gh-pages` but "do it ourselves".

From `pr-build-docs.yml`, a `keep_files` stanza is removed. With a
specified `destination_dir`, files outside that directory are already
kept as-is. There is no reason to keep files inside the directory as
well.