Correct information about RAM
An asynchronous RAM is asynchronous. A block RAM is synchronous and
pipelined.

How on earth did this happen? Of course asynchronous RAM is not
synchronous, duh. I love it, "block RAM also has a synchronous read port
and is used the same way as async RAM". Must have been some interesting
editing that lead to this.