Fix tutorial test bench
The test bench in the tutorial didn't compile, because it used
`outputVerifier` instead of `outputVerifier'` (note prime at end).
Furthermore, several engineers have commented over the years that this
structure we've always suggested in our tutorial material, with
`expectOutput` being a function that takes an entity, is a bit odd. I
feel it makes more sense to define `out` and `expected` as the signal
from the DUT and the expected signal from the DUT.