Inline comments about fixes, improvements
DO NOT MERGE to main

These are things that stood out to me while porting the text to this
repository or quickly reading through.

These HTML comments also end up in the rendered pages, but they don't
affect how the page looks in a browser. In order to maintain the
structure of the documents, the opening tag needs to be indented at
least four spaces from the line before. I did find out that if you do
this immediately after a fenced code block, it is incorporated into the
code block, and instead you should /not/ indent the opening tag.

Anyway, the comments aren't intended to end up in rendered pages, they
are just a way of annotating the source material with notes and share
those via the git repository. Please don't merge this to main :-).